[AXON-31] Extract CheckoutHelper interface
to decouple the vscode-api-heavy BB logic from URI handler
